Update README.md
This Netflix Clone project is a responsive web application built entirely using HTML and CSS to replicate the visual design and layout of Netflix, the popular streaming service. The goal of this project is to enhance web development skills by closely imitating a real-world application interface, focusing on layout structuring, styling, and responsiveness. It allows users to navigate through various categories of movies and TV shows, presenting content in a clean, modern, and visually appealing manner.